segment_tree

+++ include/structure/math/fenwick_tree.h
@@ -14,7 +14,7 @@ struct fenwick_tree {
 	std::vector<T> v;
 	std::function<T(T,T)> f;
 
-	fenwick_tree(int sz, std::function<T(T, T)> f) : v(sz + 1), f(f) {
+	fenwick_tree(size_t sz, std::function<T(T, T)> f) : v(sz + 1), f(f) {
 	}
 
 	void add(size_t index, const T& val) {
